move calculateUpdatedState to utilityfunctions
