+++ plugins/ros/src/components/riScInfo/riScStatus/RiScStatusComponent.tsx
@@ -25,7 +25,10 @@ import UpdatedIcon from './icons/updated.svg';
 import LittleOutdatedIcon from './icons/little_outdated.svg';
 import OutdatedIcon from './icons/outdated.svg';
 import VeryOutdatedIcon from './icons/very_outdated.svg';
-import { calculateDaysSince } from '../../../utils/utilityfunctions';
+import {
+  calculateDaysSince,
+  calculateUpdatedState,
+} from '../../../utils/utilityfunctions';
 
 const emptyDifferenceFetchState: DifferenceFetchState = {
   differenceState: {
@@ -168,45 +171,6 @@ export const RiScStatusComponent = ({
   const numOfCommitsBehind = selectedRiSc.numOfGeneralCommitsBehind
     ? selectedRiSc.numOfGeneralCommitsBehind.toString()
     : null;
-
-  const calculateUpdatedState = (
-    daysSinceLastModified: string | null,
-    numOfCommitsBehind: string | null,
-  ): string => {
-    if (daysSinceLastModified && numOfCommitsBehind) {
-      const days = parseInt(daysSinceLastModified);
-      const commits = parseInt(numOfCommitsBehind);
-
-      if (commits > 50) {
-        return 'very_outdated';
-      } else if (commits >= 26 && commits <= 50) {
-        if (days <= 30) {
-          return 'little_outdated';
-        } else if (days >= 31 && days <= 90) {
-          return 'outdated';
-        } else {
-          return 'very_outdated';
-        }
-      } else if (commits >= 11 && commits <= 25) {
-        if (days <= 30) {
-          return 'updated';
-        } else if (days >= 31 && days <= 90) {
-          return 'little_outdated';
-        } else if (days >= 91 && days <= 180) {
-          return 'outdated';
-        } else {
-          return 'very_outdated';
-        }
-      } else if (commits <= 10) {
-        if (days <= 60) {
-          return 'updated';
-        } else {
-          return 'little_outdated';
-        }
-      }
-    }
-    return 'updated';
-  };
 
   const updatedState = calculateUpdatedState(
     daysSinceLastModified,
